run.py tweaks
