+++ parlai/crowdsourcing/tasks/turn_annotations_static/util.py
@@ -4,15 +4,13 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-import os
 import random
-import shutil
-import subprocess
 
 from mephisto.operations.operator import Operator
 from mephisto.tools.scripts import load_db_and_process_config
 from omegaconf import DictConfig, OmegaConf
 
+from parlai.crowdsourcing.utils.frontend import build_task
 from parlai.crowdsourcing.utils.mturk import soft_block_mturk_workers
 
 
@@ -37,32 +35,3 @@ def run_static_task(cfg: DictConfig, task_directory: str):
     operator.wait_for_runs_then_shutdown(
         skip_input=True, log_rate=cfg.monitoring_log_rate
     )
-
-
-def build_task(task_directory: str):
-    """
-    Build the task with npm.
-    """
-
-    # Paths
-    frontend_source_dir = os.path.join(task_directory, "webapp")
-    frontend_build_dir = os.path.join(frontend_source_dir, "build")
-    return_dir = os.getcwd()
-
-    # Build the task
-    os.chdir(frontend_source_dir)
-    if os.path.exists(frontend_build_dir):
-        shutil.rmtree(frontend_build_dir)
-    packages_installed = subprocess.call(["npm", "install"])
-    if packages_installed != 0:
-        raise Exception(
-            "please make sure npm is installed, otherwise view "
-            "the above error for more info."
-        )
-    webpack_complete = subprocess.call(["npm", "run", "dev"])
-    if webpack_complete != 0:
-        raise RuntimeError(
-            "Webpack appears to have failed to build your "
-            "frontend. See the above error for more information."
-        )
-    os.chdir(return_dir)
